+++ src/asset_view.js
@@ -969,6 +969,11 @@ class AssetView {
             this.tree.refresh();
         }
 
+        if( this.previewPanel )
+        {
+            this.previewPanel.clear();
+        }
+
         this._processData( this.data );
     }